Revert "Add support for Fragments (#106)"
This reverts commit 5edeb6b0149ace6e46ff6f4de2a79a81c6d141e5.
