+++ README.md
@@ -268,85 +268,6 @@ final AffirmTrack affirmTrack = AffirmTrack.builder()
 Affirm.trackOrderConfirmed(MainActivity.this, trackModel());
 ```
 
-## Fragment supports
-We also support using fragment directly, only need to pass a ViewGroup id, we will put the `AffirmFragment` in this specified view.
-
-- Checkout
-```java
-    // In your activity/fragment, you need to implement Affirm.CheckoutCallbacks
-    Affirm.startCheckout(this, R.id.container, checkoutModel(), null, 10, false);
-
-    // - Affirm.CheckoutCallbacks
-    @Override
-    public void onAffirmCheckoutSuccess(@NonNull String token) {
-        Toast.makeText(this, "Checkout token: " + token, Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void onAffirmCheckoutCancelled() {
-        Toast.makeText(this, "Checkout Cancelled", Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void onAffirmCheckoutError(@Nullable String message) {
-        Toast.makeText(this, "Checkout Error: " + message, Toast.LENGTH_LONG).show();
-    }
-```
-
-- VCN checkout
-```java
-    // In your activity/fragment, you need to implement Affirm.VcnCheckoutCallbacks
-    Affirm.startCheckout(this, R.id.container, checkoutModel(), null, 10, true);
-
-    // - Affirm.VcnCheckoutCallbacks
-    @Override
-    public void onAffirmVcnCheckoutCancelled() {
-        Toast.makeText(this, "Vcn Checkout Cancelled", Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void onAffirmVcnCheckoutCancelledReason(@NonNull VcnReason vcnReason) {
-        Toast.makeText(this, "Vcn Checkout Cancelled: " + vcnReason.toString(), Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void onAffirmVcnCheckoutError(@Nullable String message) {
-        Toast.makeText(this, "Vcn Checkout Error: " + message, Toast.LENGTH_LONG).show();
-    }
-
-    @Override
-    public void onAffirmVcnCheckoutSuccess(@NonNull CardDetails cardDetails) {
-        Toast.makeText(this, "Vcn Checkout Card: " + cardDetails.toString(), Toast.LENGTH_LONG).show();
-    }
-```
-
-- Promotion
-```java
-    AffirmPromotionButton affirmPromotionButton = findViewById(R.id.promo);
-    Affirm.configureWithAmount(this, R.id.container, affirmPromotionButton, null, PromoPageType.PRODUCT, PRICE, true, null);
-```
-
-- Site modal
-```java
-    // In your activity/fragment, you need to implement Affirm.PrequalCallbacks
-    Affirm.showSiteModal(this, R.id.container, null, "5LNMQ33SEUYHLNUC");
-
-    @Override
-    public void onAffirmPrequalError(@Nullable String message) {
-        Toast.makeText(this, "Prequal Error: " + message, Toast.LENGTH_LONG).show();
-    }
-```
-- Product modal
-```java
-    // In your activity/fragment, you need to implement Affirm.PrequalCallbacks
-    Affirm.showProductModal(this, R.id.container, PRICE, null, PromoPageType.PRODUCT, null)
-
-    @Override
-    public void onAffirmPrequalError(@Nullable String message) {
-        Toast.makeText(this, "Prequal Error: " + message, Toast.LENGTH_LONG).show();
-    }
-```
-
 - Since there is no callback, it will return success after 10 seconds timeout
 - We will replace using the HTTP API after the API is done
