+++ affirm/src/main/java/com/affirm/android/AffirmClient.java
@@ -91,9 +91,8 @@ public final class AffirmClient {
                             );
                         } catch (JsonSyntaxException | IOException e) {
                             handleErrorResponse(
-                                    new APIException("Some error occurred while parsing response",
-                                            e),
-                                    listener
+                                    new APIException("Some error occurred while parsing the "
+                                            + "promo response", e), listener
                             );
                         }
                     } else {
