+++ affirm/src/main/java/com/affirm/android/CheckoutBaseActivity.java
@@ -1,17 +1,22 @@
 package com.affirm.android;
 
-import static com.affirm.android.AffirmConstants.CHECKOUT_CAAS_EXTRA;
-import static com.affirm.android.AffirmConstants.CHECKOUT_CARD_AUTH_WINDOW;
-import static com.affirm.android.AffirmConstants.CHECKOUT_EXTRA;
-
+import android.content.Intent;
 import android.os.Bundle;
+
+import com.affirm.android.model.Checkout;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
 
-import com.affirm.android.model.Checkout;
+import static com.affirm.android.Affirm.RESULT_ERROR;
+import static com.affirm.android.AffirmConstants.CHECKOUT_CAAS_EXTRA;
+import static com.affirm.android.AffirmConstants.CHECKOUT_CARD_AUTH_WINDOW;
+import static com.affirm.android.AffirmConstants.CHECKOUT_ERROR;
+import static com.affirm.android.AffirmConstants.CHECKOUT_EXTRA;
 
 abstract class CheckoutBaseActivity extends AffirmActivity {
+
+    private CheckoutRequest checkoutRequest;
 
     private Checkout checkout;
 
@@ -21,11 +26,15 @@ abstract class CheckoutBaseActivity extends AffirmActivity {
 
     abstract boolean useVCN();
 
+    abstract InnerCheckoutCallback getInnerCheckoutCallback();
+
     @Override
-    protected void onCreate(@Nullable Bundle savedInstanceState) {
-        // Hide the actionbar because the Affirm screen is already included
+    void beforeOnCreate() {
         AffirmUtils.hideActionBar(this);
-        super.onCreate(savedInstanceState);
+    }
+
+    @Override
+    void initData(@Nullable Bundle savedInstanceState) {
         if (savedInstanceState != null) {
             checkout = savedInstanceState.getParcelable(CHECKOUT_EXTRA);
             caas = savedInstanceState.getString(CHECKOUT_CAAS_EXTRA);
@@ -35,16 +44,39 @@ abstract class CheckoutBaseActivity extends AffirmActivity {
             caas = getIntent().getStringExtra(CHECKOUT_CAAS_EXTRA);
             cardAuthWindow = getIntent().getIntExtra(CHECKOUT_CARD_AUTH_WINDOW, -1);
         }
-
-        Affirm.startCheckout(this, android.R.id.content, checkout, caas, cardAuthWindow,
-                useVCN());
     }
 
     @Override
-    public void onSaveInstanceState(@NonNull Bundle outState) {
+    protected void onSaveInstanceState(Bundle outState) {
         super.onSaveInstanceState(outState);
+
         outState.putParcelable(CHECKOUT_EXTRA, checkout);
         outState.putString(CHECKOUT_CAAS_EXTRA, caas);
-        outState.putInt(CHECKOUT_CARD_AUTH_WINDOW, cardAuthWindow);
     }
+
+    @Override
+    void onAttached() {
+        checkoutRequest = new CheckoutRequest(checkout, getInnerCheckoutCallback(), caas, useVCN(),
+                cardAuthWindow);
+        checkoutRequest.create();
+    }
+
+    @Override
+    protected void onDestroy() {
+        checkoutRequest.cancel();
+        super.onDestroy();
+    }
+
+    protected void finishWithError(@NonNull Throwable error) {
+        final Intent intent = new Intent();
+        intent.putExtra(CHECKOUT_ERROR, error.toString());
+        setResult(RESULT_ERROR, intent);
+        finish();
+    }
+
+    protected void webViewCancellation() {
+        setResult(RESULT_CANCELED);
+        finish();
+    }
+
 }
