--- samples-java/src/main/java/com/affirm/samples/Config.java
@@ -1,8 +1,0 @@
-package com.affirm.samples;
-
-public class Config {
-
-    public static final String PUBLIC_KEY = "Y8CQXFF044903JC0";
-
-    public static final String MODAL_ID = "5LNMQ33SEUYHLNUC";
-}
