--- samples-java/src/main/java/com/affirm/samples/FragmentUsagesActivity.java
@@ -1,14 +1,0 @@
-package com.affirm.samples;
-
-import android.os.Bundle;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-public class FragmentUsagesActivity extends AppCompatActivity {
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_fragment_usages);
-    }
-}
