--- samples-java/src/main/res/layout/activity_fragment_usages.xml
@@ -1,4 +1,0 @@
-<fragment xmlns:android="http://schemas.android.com/apk/res/android"
-    android:name="com.affirm.samples.FragmentUsagesFragment"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent" />
