+++ samples-java/src/main/res/layout/fragment_main.xml
@@ -122,15 +122,6 @@
             android:layout_marginTop="16dp"
             android:text="@string/clear_cookies" />
 
-        <Button
-            android:id="@+id/fragmentUsages"
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:layout_below="@+id/clearCookies"
-            android:layout_centerHorizontal="true"
-            android:layout_marginTop="16dp"
-            android:text="@string/fragment_usages" />
-
         <ImageView
             android:id="@+id/imageView"
             android:layout_width="wrap_content"
