+++ samples-java/src/main/res/values/strings.xml
@@ -7,6 +7,5 @@
     <string name="track_order_confirmed">Track Order Confirmed</string>
     <string name="show_indicator">Show Indicator</string>
     <string name="clear_cookies">Clear Cookies</string>
-    <string name="fragment_usages">Fragments Usages</string>
 
 </resources>
